Update requirements and build instructions (#7)
* Update README.md

To build this project, dependencies are installed with dep. These trivial
changes documents that requirement and how to use to install dependencies
before trying to build the project.

This project requires Go v1.9 to build because of the terraform dependency it
uses.

These trivial changes replaces the generic github.com URLs originating from
the terraform plugin template, with actual URLs for this particular provider.